% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LargeScaleIncidenceRates.R
\docType{package}
\name{LargeScaleIncidenceRates}
\alias{LargeScaleIncidenceRates}
\alias{LargeScaleIncidenceRates-package}
\title{LargeScaleIncidenceRates}
\description{
LargeScaleIncidenceRates
}
